fix(separate package): Fix separate package
