+++ src/test/java/step1/racing/CarPositionViewerTest.java
@@ -8,6 +8,8 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
+import step1.racing.view.CarPositionViewer;
+
 class CarPositionViewerTest {
     @ParameterizedTest
     @MethodSource("carPositionsAndFlag")
